Separates out ZipForm, ResultsPage more into their own components (but after submitting zip form not working yet)
